desafio-complementario

+++ src/public/js/cart.js
@@ -1,35 +1,59 @@
-document.querySelector('.cancelarCompra').addEventListener('click', function() {
-    const cid = this.dataset.cid
-    fetch(`/api/carts/${cid}`, {
-        method: 'DELETE',
-    })
-    .then(response => response.json())
-    .then(data => {
-        console.log(data)
-        window.location.reload()
-    })
-    .catch(error => {
-        console.error('Error al cancelar la compra:', error)
-    })
-})
-
-document.querySelectorAll('.iconoBasura').forEach(function(button) {
-    button.addEventListener('click', function() {
-        const cid = this.dataset.cid
-        const pid = this.dataset.pid
-
-        console.log('CID:', cid)
-        console.log('PID:', pid)
-        fetch(`/api/carts/${cid}/products/${pid}`, {
-            method: 'DELETE',
+document.addEventListener('DOMContentLoaded', async () => {
+    const cartButton = document.getElementById('cart')
+        cartButton.addEventListener('click', async () => {
+            try {
+                const response = await fetch(`/api/users/user-cart`, {
+                    method: 'GET',
+                })
+                const data = await response.json()
+                const cid = data.cid
+                if (!cid) { 
+                    window.location.href = `/login`
+                    return
+                }
+                    window.location.href = `/api/carts/${cid}`
+            } catch (error) {
+                console.error(error)
+            }
         })
-        .then(response => response.json())
-        .then(data => {
-            console.log(data)
-            window.location.reload()
+    
+
+    const cancelarCompraButton = document.querySelector('.cancelarCompra')
+    if (cancelarCompraButton) {
+        cancelarCompraButton.addEventListener('click', function() {
+            const cid = this.dataset.cid
+            fetch(`/api/carts/${cid}`, {
+                method: 'DELETE',
+            })
+            .then(response => response.json())
+            .then(data => {
+                console.log(data)
+                window.location.reload()
+            })
+            .catch(error => {
+                console.error('Error al cancelar la compra:', error)
+            })
         })
-        .catch(error => {
-            console.error('Error al cancelar la compra:', error)
+    }
+
+    document.querySelectorAll('.iconoBasura').forEach(function(button) {
+        button.addEventListener('click', function() {
+            const cid = this.dataset.cid
+            const pid = this.dataset.pid
+
+            console.log('CID:', cid)
+            console.log('PID:', pid)
+            fetch(`/api/carts/${cid}/products/${pid}`, {
+                method: 'DELETE',
+            })
+            .then(response => response.json())
+            .then(data => {
+                console.log(data)
+                window.location.reload()
+            })
+            .catch(error => {
+                console.error('Error al cancelar la compra:', error)
+            })
         })
     })
 })
